added map/authentication
added map and authentication stuff
