+++ API/WebApplication/index.html
@@ -3,20 +3,127 @@
 <html>
 
 <head>
+    <link rel="stylesheet" media="all" href="css/main.css" />
+    <script src="Scripts/modernizr.custom.37797.js"></script> 
     <!-- Grab Google CDN's jQuery. fall back to local if necessary -->
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.3/js/select2.min.js"></script>
     <link href="https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.3/css/select2.min.css" rel="stylesheet" />
+    <script type="text/javascript" src="https://maps.googleapis.com/maps/api/js?key=AIzaSyCp_BKWbC2xEMhc6ARohpH7g-HV2mXZUE8"></script>
+    <script type="text/javascript" src="Scripts/jquery.googlemap.js"></script>
+    
     
     <script src="app.js"></script>
 </head>
 
 <body>
-    <div class="spotify-api-data">
-        <select>
-            <option value="3620194" selected="selected">select2/select2</option>
-        </select>
+
+    <div id="wrapper">
+
+        <header id="branding">
+            <h1>The history of aeronautics</h1>
+        </header>
+
+        <nav id="primary">
+            <ul>
+                <li>
+                    <h1>First manned flights</h1>
+                    <a class="manned-flight" href="#manned-flight">View</a>
+                </li>
+                <li>
+                    <h1>The frameless parachute</h1>
+                    <a class="frameless-parachute" href="#frameless-parachute">View</a>
+                </li>
+                <li>
+                    <h1>Over the English Channel</h1>
+                    <a class="english-channel" href="#english-channel">View</a>
+                </li>
+                <li>
+                    <h1>About</h1>
+                    <a class="about" href="#about">View</a>
+                </li>
+            </ul>
+        </nav>
+
+        <div id="content">
+            <article id="manned-flight">
+                <header>
+                    <h1>First manned flights</h1>
+                </header>
+                <div class="spotify-api-data">
+                    <select>
+                        <option value="3620194" selected="selected">select2/select2</option>
+                    </select>
+                </div>
+                <nav class="next-prev">
+                    <hr />
+                    <a class="next frameless-parachute" href="#frameless-parachute">Next</a>
+                </nav>
+            </article>
+
+            <article id="frameless-parachute">
+                <header>
+                    <h1>The frameless parachute</h1>
+                </header>
+                <p>André-Jacques Garnerin carried out the first jump with a silk parachute on October 22, 1797 at Parc Monceau, Paris. Garnerin's first parachute resembled a closed umbrella before he ascended; at a height of approximately 3,000 feet he severed the rope that connected his parachute to the balloon, and the basket fell to earth. Garnerin emerged uninjured.</p>
+                <nav class="next-prev">
+                    <a class="prev manned-flight" href="#manned-flight">Prev</a>
+                    <hr />
+                    <a class="next english-channel" href="#english-channel">Next</a>
+                </nav>
+            </article>
+
+            <article id="english-channel">
+                <header>
+                    <h1>Over the English Channel</h1>
+                </header>
+                <div id="map" style="width: 300px; height: 300px;"></div>
+                <nav class="next-prev">
+                    <a class="prev frameless-parachute" href="#frameless-parachute">Prev</a>
+                    <hr />
+                    <a class="next about" href="#about">Next</a>
+                </nav>
+            </article>
+
+            <article id="about">
+                <header>
+                    <h1>About</h1>
+                </header>
+                <p>This is a jQuery parallax scrolling experiment by Jonathan Nicol. <a href="http://f6design.com/journal/2011/08/06/build-a-parallel-scrolling-website-interface-with-jquery-and-css/">Read about it</a> on my blog.</p>
+                <nav class="next-prev">
+                    <a class="prev english-channel" href="#english-channel">Prev</a>
+                    <hr />
+                </nav>
+            </article>
+        </div>
+
+        <!-- Parallax foreground -->
+        <div id="parallax-bg3">
+            <img id="bg3-1" src="img/balloon.png" width="529" height="757" alt="Montgolfier hot air balloon" />
+            <img id="bg3-2" src="img/balloon2.png" width="603" height="583" alt="Frameless parachute" />
+            <img id="bg3-3" src="img/balloon3.png" width="446" height="713" alt="Blanchard's air balloon" />
+            <img id="bg3-4" src="img/ground.png" width="1104" height="684" alt="Landscape with trees and cows" />
+        </div>
+
+        <!-- Parallax  midground clouds -->
+        <div id="parallax-bg2">
+            <img id="bg2-1" src="img/cloud-lg1.png" alt="cloud" />
+            <img id="bg2-2" src="img/cloud-lg1.png" alt="cloud" />
+            <img id="bg2-3" src="img/cloud-lg1.png" alt="cloud" />
+            <img id="bg2-4" src="img/cloud-lg1.png" alt="cloud" />
+            <img id="bg2-5" src="img/cloud-lg1.png" alt="cloud" />
+        </div>
+
+        <!-- Parallax  background clouds -->
+        <div id="parallax-bg1">
+            <img id="bg1-1" src="img/cloud-lg2.png" alt="cloud" />
+            <img id="bg1-2" src="img/cloud-lg2.png" alt="cloud" />
+            <img id="bg1-3" src="img/cloud-lg2.png" alt="cloud" />
+            <img id="bg1-4" src="img/cloud-lg2.png" alt="cloud" />
+        </div>
+
     </div>
+
 </body>
 
 </html>
